Make a copy of an existing DeliteFileInputStream in readLinesChunk.
This allows us to avoid re-indexing all the files in a stream (i.e.
re-reading all the file descriptors) as we chunk through it.
Furthermore, we also now require a newly constructed stream to be
explicitly opened, as the implicit open before caused streams
constructed in readLinesChunk to be opened but never closed.